Move the global AST vectors to their own header
This is a refactoring step in eliminating the use of globals
throughout the compiler, with the first target being passes.

As passes are rewritten to not use these globals, the header include
line can be moved down below any new implementation to ensure no
accidental usage is introduced. This makes it seamless to replace
incrementally so that the existing pass function can still access
globals, but we don't have to split a new TU to get the isolation we want

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>