Introduce PassManager and use in returnStarTuplesByRefArgs
We add a basic new class called PassManager and its associated PassT
class to begin adding structure to the passes we have. Passes
implemented as these classes are intended to be isolated from its
environment so that we can start decoupling when passes get run and on
which parts of the AST. Today, things are not so isolated so we need
to get their incrementally. The ComputeCallSites pass for example
looks up some info about virtual methods in dispatch.cpp, so that
loose end needs to get chased down in the future.

As a first step, the PassManager supports calling one or more passes
that all operate on the same type of input. It then takes care of
calling each pass on each object, which can be done in either loop
order (see `runPass` vs `runPassChained`). As mentioned in the
comments, we could build larger trees/graphs of passes and do the same
thing, but the easiest way to interoperate (and have the same
performance) with the existing compiler is to maintain these passes
over things of the same type

A PassT is informed by patterns we use, where the dominant one is
"conditinally do this thing over all the things of type __". The run
method mimics an operator() and defaults to a noop pass that processes
everything, but does nothing. Note that we have two different forms of
"query" here, one is for type and one is reflected in `shouldProcess`
b/c even within an AST node type, there are still more properties that
determine whether we should run. One possible direction to go with
this now that we've given it a name is to use it for an idempotence
check (so that a demand-based approach could be taken).

These new passes are currently listed in passes.h This was the easiest
place to drop them for now but we could consider splitting them out.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>