Don't slacken the MR extent test if the comm cache is disabled.
If the comm cache is not in use there is no need to add any slack to the new
MR extent checks, because it's the cache's cacheline (pre)fetching that caused
the need for that slackening in the first place.  Therefore, don't slacken the
check when the cache is not in use, thus keeping the check as tight as we can
in all situations.

I tested this by building and running the `miniUnorderedCopyStress` test both
with and without the comm cache to make sure it worked in both cases, and then
temporarily reversing the sense of the `chpl_cache_enabled()` test in the code
and building and running `miniUnorderedCopyStress` with the cache enabled.  I
expected that doing so would cause an (incorrect!) MR extent failure, and
indeed it did.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>