Add comments for why we disable merging/splitting under ugni
In 18738 we disabled merge/split hooks under ugni as a workaround to a
bug we didn't fully understand. We have since figured out that the
problem was that we had a single blob of memory spanning multiple memory
registration regions created by merging chunks, and this isn't something
ugni supports. 19127 added checks for this condition and this just
adds comments to indicate why we disable merge/split for ugni.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>