Update .good files to allow testing with network atomics.
The `.good` files for `promotion` and `promotionMany` only matched in
the absence of network atomics support.  With `CHPL_COMM=ugni` and `ofi`
they did not match, because the things done with execute-on-fast comm
ops are instead done with (unstable) "network" atomics.  Here adjust the
`.good` files so that these tests pass with or without network atomics.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>