Deprecate collect_asts
`collect_asts` is currently used in multiple situations, both where
the caller depends on the order being a preorder, and where the
ordering doesn't matter. This mostly informational change separates
those two into collect_asts_{un,pre}order so that the caller's intent
can be specified.

This change is motivated by wanting the freedom from ordering in order
to speed up the unordered case. I only discovered that at least one
callsite relied on the order being a preorder by attempting to make an
impl change and getting an error.

My initial testing shows that the one callsite modified in
scopeResolve is the only caller that relied on the order being a
preorder, but this isn't conclusive and so the remaining callsites of
`collect_asts` should remain as "assume they require preorder, but
probably not and switch them if that is the case after careful
reasoning about the code".

This change is also motivated by the ongoing work of introducing a
PassManager, which will be assuming some of the responsibility of iterating
over some child nodes, and it is very useful to know which callsites
need what behavior so we can safely modify them. It also is relevant when
thinking about adding some other method to quickly iterate over a
subset of nodes (like a per-node type linked list).

As noted, this same problem exists with some of the other collect*
functions and I hope we can transition the others in a similar
manner.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>