Remove deprecated oldCollectors.h and add iterator adapters
Removes collect_asts_postorder given a Vec in favor of the one taking
a std::vector

Vec -> std::vector required add -> push_back and forv_Vec to a range
for. I opted for range for over for_vector because it is standard for
c++14. One loop needed to be rewritten as a for i loop because it is a
known expanding loop. Other expanding loops are marked in the next commit

In a similar spirit, we also move the stlUtil.h macros one step
towards removal by adding iterator adapters that maintain the slightly
unusual nature of their predecessor (which is probably accidental in
most uses).

I've left removing the stulUtil.h macros for future work as I think
they should either be addressed individually or tackled as a whole,
but with more time and effort than I have at this moment.

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>