Add github CI runner for dyno linter
This adds a github CI action to run the target run-dyno-linters

A sample runtime of the linters on github CI was 1m47s for the actual
run step

I've left the target executable as compiler/dyno/util/lint so it can
serve as an entrypoint and it takes care of running the tool on the
proper files.

I've enabled generating the compile_commands.json compilation database
from within cmake so that the fieldsUsed linter can use it

In the future, we could consider integrating the lint running into
cmake itself, but I think this approach fits better with how we
currently have things setup. (Plus I'm not sure how to integrate
custom tools into the builtin lint checks like CMAKE_CXX_CLANG_TIDY)

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>