Move CPtr's contents into CTypes
As part of this, I associated all operators with one type or another.
In some cases I think these choices are completely reasonable (e.g.,
since `int` is always available, an operator between 'c_ptr' and int
makes sense to associate with 'c_ptr' in this module).  In other
cases, it was fairly arbitrary so that I could publicly use them from
CPtr.  Since they're not documented, it seems as though this isn't
currently a user-facing change.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>