Don't run n=10000 problem size in valgrind
In the "Brad can be a slow learner" category, this moves the .execopts
that I applied to my study version of pidigits (in #19189 to avoid a
valgrind timeout) over to the snapshot of the submitted version.
Since my version is now simple / serial, I also restored the
traditional .execopts for my study version.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>