Refactor of localizeGlobals to use PassManager
This refactors the localizeGlobals pass to use PassManager with the pass
  LocalizeGlobals

Here we keep some collections around -- `globals` and `symExprs` for
re-use. They are manually reset with `clear` at the beginning of
`process`, which I've pointed out might be a pattern worth considering
adding to PassT, but we'll see.

I also identified the pattern
  collect_asts asts
  for ast in asts:
    if ast is SymExpr:
      ...
which I've switched to use collectSymExprs which does this filtering
itself. I'm also marking loops that don't modify things with TODO
TRAVERSE as they are good candidates for using a traversal in the
future instead of a collect and iterate approach

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>