Refactor of bulkCopyRecords with PassManager
This refactors the bulkCopyRecords.cpp pass with the pass
  BulkCopyRecords

This one especially has some utility functions that I've pulled in as
static methods so that they could presumably be tested in the
future. I'm finding that many passes define predicates like
`isAssignment` and I'm not immediately jumping to centralizing them
b/c a) I don't know if they would get reused and b) I suspect what
counts as an assignment to one pass may not be the same as before. I
think we can take a pass later over all these static methods and
decide what can be lifted out into a standalone. Note that one
complication is that certain pure methods use a map for efficiency or
termination. In this pass, `typeContainsRef` is an example of that
which is an instance method so it can access the `containsRef` map

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>