Refactor of removeUnnecessaryAutoCopyCalls with PassManager
This refactors the removeUnnecessaryAutoCopyCalls pass into
  RemoveUnnecessaryAutoCopyCalls

As noted in a comment, this one calls into a
removeInitOrAutoCopyPostResolution from
propagateDomainCorrectness. I think this is a pure helper and seems
like it would be good to have a category (in some c++ structured way
whether static methods or in a namespace) of these helpers that
operate on a particular ast node and don't mess with global
state. Leaving this as something to keep in mind going forward

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>