Deprecation and docs adjustments
- Change numLimbs and get_limbn back to referring to the mpz field, mpz_t is not
the same type as __mpz_struct and it seemed too complicated to describe how to
use __mpz_struct in the same way.  The mpz field is not going away, it's just
getting "no doc"ed, though it may become private in the future.
- Make the mpzStruct deprecation message list its name, to be more helpful in
case multiple deprecated functions are used in the same line
- Reword the documentation for getImpl a little bit (but not much)

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>