Rename 'off_t'/'mode_t'/'socklen_t' to use c_ prefix
This deprecates the types:

* off_t
* mode_t
* socklen_t

replacing them with c_off_t, c_mode_t, and c_socklen_t as proposed
in #18012.  This has the beneficial side-effect of resolving an
issue that I introduced in #19262 in which my trick for forwarding
these types from SysBasic to Sys caused problems when both modules
were 'use'd because it had the effect of bringing a variable and
procedure overload into the same scope which the compiler doesn't
permit.

Update uses of these routines in tests and modules, and replace
the forwarding procedures with deprecated types now that we're
using different types.  Add a new test that captures the case
that had not worked properly before.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>