Manually disable catch-all min()/max() overloads for arrays
These overloads don't work when promoted due to the lack of promotion
support in conditionals, but if they're not used for arrays, things
work as you'd expect (by promoting the scalar versions).  So disable
the array case to permit elementwise min/max on arrays.

This makes me wonder a bit about what more we could do as
language/library designers to make such catch-all overloads more
constrained; or what practices we should use as library developers
to avoid such cases.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>

Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>