Restore intended PTRANS.valgrind behavior
This test's behavior started changing when we started running
initialization for all modules named on the command-line because we
were calling main once for the `--main-module` of PTRANS and once for
the explicit call to main() in the PTRANS.valgrind module.  When
writing this, the author (could've been me) presumably imagined that
their explicit call to main() was happening and necessary when in fact
it wasn't.

Thanks to Andrew for helping to diagnose this.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>