Change deprecation warning for operators declared with `proc` to just an error
This keeps the parser relatively simple while still giving a good error message
when trying to declare operators incorrectly.  Also, don't worry about if it's
declared as a method or not, it should use the keyword regardless

Move the deprecation test into the main test hierarchy for operators, but
changes its expected output to match the new compilation failure.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>