Test updates for changing the operators declared without the keyword to failures
These tests all used a prediff that effectively suppressed the deprecation
warning, so we didn't notice they needed updates before.  Now that the syntax
results in a compilation failure, the test system flags them.

Includes adding a .notest to the submitted mandelbrot3 since updates to that
directory would not match the version on the shootout site.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>