Answer a comment Question during refactor
I had left Q in passing and have answered it with the help of
dlongnecker

This function looks in PrintModuleInitOrder.chpl for variable with the
pragma mentioned. It would be an error to have more than one of these
and it would also be an error to have zero of them because downstream
passes rely on gModuleInitIndentLevel to be non-null

We could arguably add the break I had mentioned in the comment, but
the number of symbols is small for that module and it is nice to have
the sanity check that we don't find more than one.

Tested by removing and then also by duplicating the var in that module
and observed an error as expected. Normal compiles unaffected

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>