Remove remaining deprecated Path module methods on the file type
These were deprecated in 1.25 in favor of standalone functions on the file type.
Remove them now that the module is stable and users have had a release cycle to
notice the deprecation warning.

While here, noticed that the error messages in `realPath(f: file)` still
pretended to be in the method version, so updated their wording.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>