Remove previously deprecated instances of regexp
Removes:
- initializer in RecordParser module
- createRegexp function in RecordParser module
- CHPL_REGEXP module-scoped variable in ChplConfig module
- old Regexp module
- regexp type in new Regex module
- checks involving CHPL_REGEXP in util/chplenv/chpl_re2.py

Updates
- test/chplenv/chplconfig/warnings/chplconfig and the test file that uses it to
no longer check for the environment variable warning
- test/deprecated/regexp_error.chpl to use the regex type instead of regexp

Removes tests of the deprecated symbol in ChplConfig (previously ChapelEnv) and
the deprecated Regexp module.

Note: does not remove BadRegexpError - due to a bug, this was originally not
generating a deprecation message.  I think it can now (though by using the
deprecated keyword instead of the function that encounters the bug), so I
think it should be updated to do so (and will in a later commit)

Does not remove `ok` and `error` methods.  I think one could argue that those
fall under the "not ready to be cleaned up yet" category.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>