gModuleInitIndentLevel only needs to exist for non minimal-modules
In #19273 I tacked on an additional commit that strengthened an
assert, but this did not apply for minimal modules. Because this was
an tack-on commit and I was in a rush, something in my testing run did
not work properly.

The proper behavior is that we only expect the
PrintModuleInitOrder module to exist if not in minimal modules mode

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>