Use deprecated keyword with BadRegexpError
We were originally stymied from adding a deprecation warning for this symbol due
to a bug with returning types from functions and using the result in catch
clauses.  Now that the deprecation keyword can be applied to any symbol, it's
possible to deprecate a type alias, so do so.

Unfortunately, this encounters a bug with deprecating symbols used in catch
clauses, where such symbols will result in multiple deprecation warnings for
a single reference - however, two messages is better than none and will allow
us to move towards removing this symbol in the meanwhile.

This commit adds the deprecation warning and adjusts the already existing test
of other error deprecations to use the deprecated Error subtype.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>