Fix sporadic gpu issue
Do not require GPU context to be initialized prior to calling functions
that call out to chpl_gpu_mem_alloc.  chpl_gpu_mem_alloc() already
checks to see if the context has been initialized and will initialize it
if needed.

This caused a sporadic issue (race condition) where sometimes the memory
allocation function would be called before something else caused the
context to be initialized. In such cases, the memory allocation would be
done on the CPU (and we'd segfault after a GPU kernel tries to access
that memory)

---
Signed-off-by: Andy Stone <stonea@users.noreply.github.com>

Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>