Revert Reverting #18725 to allow calling some functions from GPU kernels
We previously reverted #18725 which interfered with another PR that was added
later. Add #18725 back with modifications to get it testing cleanly.

Instead of issuing an error for a function called from a gpu kernel that
accesses outer variables, decline to translate it into a gpu kernel in the
first place.

Update a test to stop expecting the error. Adjust a PREDIFF to strip out
some expected bad output.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>