Wait for catfiles to exist in sub_test
Wait up to 10 seconds for catfiles to exist. Catfiles are used to add
files generated by tests to execution output in order to diff them
against .good files. On systems with a network mounted filesystem and a
launcher we have seen cases where the launcher returns before the file
is visible on disk. This adds up to a 10 second delay waiting for
catfiles to exist. In practice it's fractions of a second delay that we
see for our tests, but this seemed a little more robust.

This is motivated by testing failures we've seen on some Cray systems
when launching with `srun`. Note that this also changes a manual cat in
a prediff to use a catfile and adds a sleep to a test prediff that does
something like catfiles but more involved with sorting and diffing.

Part of Cray/chapel-private 3147

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>