Don't warn when `CHPL_TARGET_CPU` is set to the inferred value on Crays
On Cray XC/EX `CHPL_TARGET_CPU` is based on the loaded `craype-` cpu
targeting module and we don't want users to set it explicitly so we warn
if they do. However, there's no reason to warn if the user set the value
to the inferred one so stop doing that here. This was causing test
failures for a prediff because we inject the chplenv into the prediff
env, which resulted in `CHPL_TARGET_CPU` being set explicitly, which
caused the printchplenv called from the prediff to warn. While here fix
a python2/3 incompatibly in the prediff.

Part of Cray/chapel-private 3147

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>