Refactor insertLineNumbers to use the PassManager
This is a refactor of the insertLineNumbers (and insertNilChecks)
pass(es) to use the PassManager

insertNilChecks is straightforward and all good -- though perhaps out
of place for this pass

The real crux in insertLineNumbers is that it quickly pops out of
isolation. With the PassManager, we'd like to move towards a world
where a pass operates on a Function (for instance) and only modify
things it contains. In this case, we run into a scenario where we need
to modify ourselves (the Function we're working on) AND all of its
call sites. This is a common pattern in whole program compilers.

To do that, the InsertLineNumbers has a (unordered) queue of Functions
that need to be processed. When processing our own function, if we
need to modify someone else, we instead enqueue the work and trust
that the PassManager will drain the queue. Similarly, when we need to
process a Call that does not live in the currently processed Function,
we instead enqueue it because processing it may modify its enclosing
Function

By queueing work later and waiting for the PassManager to drive it,
we're giving the PassManager a place to -- in the future -- mediate
these modifications so they may take place on different parts of the
AST in different parts of the compilation pipeline

In this refactor, I've also removed the need for "move line inside
args bundle" because I'm handling them in the general case instead of
fixing them up afterwards

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>