Use hugepages on the hpe-cray-ex platform.
Enable use of hugepages on the hpe-cray-ex platform. Due to limitations in the
hugepage interface, the runtime will not size the heap dynamically based on
available physical memory, but will instead use the default heap size or the
size specified by CHPL_RT_MAX_HEAP_SIZE. The program may fail with an
out-of-memory error if there is insufficient physical memory for the heap.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>