Test changes for '--dry-run'.
Adjust testing to reflect changes due to the addition of '--dry-run'.  A
major difference is that now with all launchers, '--help' will produce
the "LAUNCHER FLAGS:" secion of the output. Before, we didn't get that
when using launchers such as `amudprun` that lack any launcher-specific
flags.  In several tests that are really just about config var handling
I did that by adjusting prediff files to discard altogether the parts of
'--help' output that don't have to do with config vars.

A side effect of the change in how we print '--help' output was the
removal of a long-bothersome double space in `-v` and `--verbose`
output.  There are small changes here to testVFlag and testVerboseFlag
`.good` files to adjust to that.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>