Improve oversubscribedArrayAlloc test for configs using a fixed heap
This test attempts to use a large fraction of memory to stress test
array allocations. However, it was trying to use a fraction of physical
memory, which can be much larger than what we can actually allocate when
using a fixed heap with `CHPL_RT_MAX_HEAP_SIZE` set lower, which we do
sometimes to speed up testing.

Base the amount of memory used on the heap size if one is used to avoid
this issue.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>