Add a test to stress different communication operations
Add a test / micro-benchmark that stresses individual communication
operations (GETs, PUTs, fetching and non-fetching AMOs, AMs). This has
been useful for exploring comm layer performance so I wanted to get it
committed. For now it's only run as a correctness test, but I expect to
make it a performance test soon.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>