Improve comm-ops test
Previously, the comm-ops test  was doing comm on Locale 1, to/from
Locale 0. This was an artifact of how the test was originally written
with a default rectangular array on locale 0, but now that it's using a
blockdist array there's no reason to do that.

While here update the test to grab references to array elements to avoid
repeated blockdist indexing. This gives us a more true sense of the
overhead of just the runtime functions and ignores indexing overhead.

Also add support for testing fast vs non-fast AMs

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: slnguyen <26969019+slnguyen@users.noreply.github.com>