First attempt to set up Makefile to enable 'make -j' with '--incremental'
Because, in the past, we used a 'foreach' in our Makefile to compile
all of the component object files, our generated Makefile was not
amenable to using 'make -j'.  This PR restructures things to list
all .o files as dependencies of the binary, permitting them to
be parallelized with 'make -j' (I believe).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>