Fix a domain equality comparison in SSCA2
SSCA2 used to compare domains of different kind, which is illegal
as of #19356. This adds a guard that enables the comparison only
when the two domains are of the same kind.

Digging deeper, there are two uses of the enclosing function:
when the two domains are distinct and when they are the same domain.
So perhaps the domain comparison is used to distinguish these two uses.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>