Adjust the madness test directory skipif so it only excludes testing with intel
Some of the tests in this directory are sensitive to numerical differences
between compilers.  Through experimentation, it seems like the gnu and Cray
compilers match the behavior of LLVM, our new default, while intel has slightly
different behavior.

We want to run these tests in our default settings so that problems with them
can be detected early, so an update to the skipif was necessary.  Since these
other configurations match the behavior of our default, at least for now test
with them as well.

Note: it might be nice to have alternative .good files for intel, but we're
worried that this would result in unnecessary overhead to maintain, so for now,
just skip it.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>