Add testing for llvm 11 and 12
Now that we support multiple versions of llvm, add some explicit testing
for older versions. The current plan is to default to and recommend 13,
but we want some level of testing on 11 and 12 still so add that here.
Currently this just tests `release/examples` for 11 and 12 on x86 with
comm none, but we may want to add more coverage in the future (more
tests, add gasnet, check on arm too.)

Part of Cray/chapel-private 1031

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>