Optimize `myLocaleSpace` assignment in aggregators
Change `myLocaleSpace = LocaleSpace` to `myLocaleSpace = 0..<numLocales`
to eliminate communication when creating aggregators. `LocaleSpace`
lives on locale 0, so there was some communication to copy it. Using
`numLocales` eliminates that because `numLocales` is just an `int` that
is replicated at program startup. This eliminates a many-to-one GET,
which may have a minor performance benefit, particularly at scale.

While here, also change the non-ugni default buffer size from `8096` to
`8192`. I meant for this to be a power of two in 18326 but either made a
typo or a think-o.

Note that this is a clone of changes made in Bears-R-Us/arkouda 1082

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>