Migrate copy aggregators to a package module
We've long wanted to make aggregators user-facing (16963), but haven't
made much progress on that. Longer term we want better names, support
for third-party operations, arbitrary user-defined operations, and
probably some other things, but in the short term this just exposes the
implementation for users to try out.

Note that this is a straight code move, docs will be added next.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>