Workaround sporadic cygwin compilation errors
On cygwin we've been seeing an issue where after a test is run,
some other process has a handle on the executable and we can't create a new
executable with the same name for a bit. If a test ran and we have
additional compopts (which means the exec name will be reused) wait a
second while cleaning up the executable to give the other process time
to release its handle.

See Cray/chapel-private 3143 for more info

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>