Update good file and add bad file for const-ref-tuple-decl test
The good file had gone stale wrt the arrays in main() no longer being
separately default inited and then assigned to.  They're now
initialized directly.  Update the .good file by removing the default
init of the array elements, the assignments, and the deinit of the
temporaries the array had been initialized from.

Add a .bad file.  This required adding a serial block around the body
of commonPrefixLength().  The undesired array copying was parallel and
nondeterministic.  Add a serial block around it all so that the output
is deterministic.  This does not appear to affect the behavior of the
code under test or of the correctly-functioning alternative.

Add a note to remove the serial block once the future is retired.

---
Signed-off-by: Paul Cassella <gitgitgit@manetheren.bigw.org>