Discontinue use of the fallback "system" libfabric install.
We've been maintaining a fallback "system" libfabric install from `css`,
which the HPE Cray EX module build was using if `pkg-config` didn't
already know about a real one.  But that fallback was only used when we
were building the EX module on an XC (early days!).  Now we always build
it in a virtual environment which doesn't even mount that `css` path.
Therefore here, get rid of all the logic associated with the fallback
"system" libfabric install.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>