Make --incremental even more fine-grain
The --incremental compiler flag generates a distinct C file per *user*
module, but standard and internal modules still get rolled
(conceptually, via #includes) into a single _main.c file.  The result
is that the compilation of the _main.c file tends to be a bottleneck
relative to the other ones.

This PR changes this so that every module is compiled as its own
distinct C file.

While here, I also moved the user error for compiling with
target-compiler=llvm and --incremental earlier in the compiler
so that time isn't wasted before hitting it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>