Add support for a developer '--parallel-make' flag and enable by default
This adds support for a '--parallel-make' / '-j' flag that will use
'make -j' to build the back-end C code.  It is set up such that:

* a value < 0 indicates it is disabled
* a value == 0 indicates to use 'make -j'
* a value > 0 indicates to use 'make -j <val>'

It is set up so that if we're not compiling with --fast, it will use
'0' by default, and so that if the value is >= 0, it will enable
'--incremental' by default as well.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>