Switch test-memLog to use a catfile
Previously, the test was manually looking at some memory log files, but
recently on some systems files generated on compute nodes aren't visible
to the login node for a bit. We added a wait for catfiles to work around
this, so just switch this test to use a catfile instead of manually
opening the files.

Part of Cray/chapel-private 3147

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>