Have auto-aggregation call aggregator implementation directly
In 19386 I added a wrapper to avoid aggregating for comm=none and to
make documenting easier. This was causing some compiler failures for
`cleanupRemainingAggCondStmts()` because the AST now had an additional
wrapper. I didn't feel comfortable updating this code, so for now I just
updated the code to call the non-wrapped aggregator directly.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>