Add container for Github CI actions
This adds a Dockerfile for use with our CI actions to reduce their
execution time by eliminating the need to download dependencies.

This is just phase 1 of 2, where we build the dockerfile and push it
to ghcr (github container registry). The second phase is to actually
use it in CI.yml -- I just don't know how to order these actions
correctly so I'm doing in two PRs

The container image should only get rebuilt when its Dockerfile
changes -- in other words, 99% of PRs shouldn't run this action

Signed-off-by: Andrew Consroe <andrew.consroe@hpe.com>